+++ src/components/feature/SoundStatus.tsx
