+++ src/components/tables/tts/TTSPlaybackHistory.tsx
@@ -2,8 +2,8 @@ import { Checkbox } from '@radix-ui/react-checkbox';
 import { useEffect, useMemo, useState } from 'react';
 import { TbX } from 'react-icons/tb';
 
-import { AudioPlayer, PlayerMode } from '@/components/audio/AudioPlayer';
-import { SoundStatus, UNIT_SOUND_STATUS_TYPES } from '@/components/audio/SoundStatus';
+import { AudioPlayer, PlayerMode } from '@/components/feature/AudioPlayer';
+import { SoundStatus, UNIT_SOUND_STATUS_TYPES } from '@/components/feature/SoundStatus';
 
 interface TTSPlaybackHistoryProps {
   id: string;
